fix: add missing request handler for rss feeds
