+++ packages/cron-jobs/package.json
@@ -22,7 +22,6 @@
   },
   "prettier": "@homarr/prettier-config",
   "dependencies": {
-    "@extractus/feed-extractor": "^7.1.3",
     "@homarr/analytics": "workspace:^0.1.0",
     "@homarr/auth": "workspace:^0.1.0",
     "@homarr/common": "workspace:^0.1.0",
