+++ packages/definitions/src/docs/homarr-docs-sitemap.ts
@@ -165,6 +165,7 @@ export type HomarrDocumentationPath =
   | "/docs/getting-started/installation/home-assistant"
   | "/docs/getting-started/installation/portainer"
   | "/docs/getting-started/installation/qnap"
+  | "/docs/getting-started/installation/railway"
   | "/docs/getting-started/installation/saltbox"
   | "/docs/getting-started/installation/source"
   | "/docs/getting-started/installation/synology"
