+++ packages/request-handler/package.json
@@ -22,6 +22,7 @@
   },
   "prettier": "@homarr/prettier-config",
   "dependencies": {
+    "@extractus/feed-extractor": "7.1.3",
     "@homarr/common": "workspace:^0.1.0",
     "@homarr/db": "workspace:^0.1.0",
     "@homarr/definitions": "workspace:^0.1.0",
